Move averaging logic to SDSyncAveragingService
