--- AirCasting/SDSync/Models/SDSession.swift
@@ -1,9 +1,0 @@
-// Created by Lunar on 09/12/2021.
-//
-
-import Foundation
-
-struct SDSession: Hashable {
-    let uuid: SessionUUID
-    var lastMeasurementTime: Date?
-}
